requirements.txt
requests>=2.31.0
tiktoken>=0.5.0
matplotlib>=3.8.0
pandas>=2.1.0
python-dotenv>=1.0.0
